Rename : dummyData 파일에 있는 type 다른 파일로 분리
- 기존 DummyData.ts에 정의되어있는 type Interface.ts로 분리
- API response 에 대한 data type 설명 주석 추가
